Refactor: 파일 구조 변경, Utils 추가
- global/schedule/batch.schedule... 이었던 것들을, domain/trials 안으로 옮기고 trialService로 이름 변경
- batch에 쓰이는 함수 중에, 재사용이 가능한 것을 Utils로 분리
